Moved config init from env to Configuration class
